requirements.txt
langchain_core==0.1.33
langchain_experimental==0.0.55
langchain_openai==0.1.0
python-dotenv==1.0.1
Requests==2.31.0
selenium==4.18.1
